Added new version of ftpcam
This is the final version of ftpcam, now dubbed cloudcam. The run.py file runs the photometry part, analyzing the .fits file from a given camera. a.py runs the matrix approach and calls run.py. The ftpcam.py file inside the ftpcam folder can be adapted to invoke a.py and run the full process in real time for any given ftp fits folder.